AnalisisForense/ExtractorDeMetadatos/requirements.txt
# Librerías necesarias para el proyecto
python-dotenv
PyPDF2
python-docx
exifread